+++ x/simulation/simulate.go
@@ -99,7 +99,7 @@ func SimulateFromSeedX(
 	startTime := time.Now()
 	logger.Info("Starting SimulateFromSeed with randomness", "time", startTime)
 	logger.Debug("Randomized simulation setup", "params", mustMarshalJSONIndent(params))
-	simTxLifecycleStats.reset()
+	ResetTxLifecycleFailuresForApp(app)
 
 	timeDiff := maxTimePerBlock - minTimePerBlock
 	accs = randAccFn(r, params.NumKeys())